[관리자] 불량개선률 integer -> double 변경

+++ service/src/main/java/com/kap/core/dto/mp/mpb/MPBBusDto.java
@@ -79,7 +79,7 @@ public class MPBBusDto extends BaseDTO {
     private String guidePscndDt;
 
     @Schema(title = "개선률", example = "5")
-    private Integer fltyImpvmRate;
+    private double fltyImpvmRate;
